+++ addons/main/config.cpp
@@ -40,7 +40,6 @@ class CfgFunctions
             class loopRegister {};
             class loopWatch {};
             class everyInterval {};
-            class offServer {};
             class requestId {};
             class prngNew {};
             class prngNext {};